[ALPS04973279] BACKPORT: kheaders: only headers into kheaders_data.tar.xz
Currently, kheaders_data.tar.xz contains some build scripts as well as
headers. None of them is needed in the header archive.

For ARCH=x86, this commit excludes the following from the archive:

  arch/x86/include/asm/Kbuild
  arch/x86/include/uapi/asm/Kbuild
  include/asm-generic/Kbuild
  include/config/auto.conf
  include/config/kernel.release
  include/config/tristate.conf
  include/uapi/asm-generic/Kbuild
  include/uapi/Kbuild
  kernel/gen_kheaders.sh

This change is actually motivated for the planned header compile-testing
because it will generate more build artifacts, which should not be
included in the archive.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Reviewed-by: Joel Fernandes (Google) <joel@joelfernandes.org>

MTK-Commit-Id: d0189316b15e105c635501624fe3350cd5ec47c5

Change-Id: I164f3537e33bd28fdfbf13599ac768caab17e93c
Signed-off-by: James Hsu <james.hsu@mediatek.com>
CR-Id: ALPS04973279
Feature: [Module]Official Kernel Patch