[ALPS04982554] thermal: add dummy header
[Error]
mtk_ta.c:14:10: fatal error: mach/mtk_thermal.h file not found

[Detail]
We only keep a specific platform when we run the release policy.
In the case the makefile set to include mt6779 platform.
That is not allowed when we only release mt6761 drivers.
This patch set a common platform header that is cloned
from mt6779 and set as default inclusion folder.

MTK-Commit-Id: cbaaef365a3f830bef9d55844ef7245226a64a2c

Test: Build and Boot
Change-Id: I7a85d1de277d15dd17bc980ca9c4502139005a9c
CR-Id: ALPS04982554
Feature: Thermal Management
Signed-off-by: Juju Sung <juju.sung@mediatek.com>