[ALPS04982554] power: rename platform header
[Error]
mtk-scpsys.c:19:10: fatal error: dt-bindings/power/mt2701-power.h
file not found

[Detail]
We only keep a specific platform when we run the release policy.
In the case scpsys.c include mt2701 headers.
That is not allowed when we only release mt6761 drivers.
This patch rename to v platform header to workaorund it.

MTK-Commit-Id: 2cf04e1ec19934f2e8d33350a6374ae8663da316

Test: Build and Boot
Change-Id: I89e924a934cdfa8efc5b948b6767f962dd2ea73d
CR-Id: ALPS04982554
Feature: Thermal Management
Signed-off-by: Juju Sung <juju.sung@mediatek.com>