[ALPS04982554] power: rename platform header
[Error]
mtk-scpsys.c:19:10: fatal error: dt-bindings/power/mt2701-power.h
file not found

[Detail]
We only keep a specific platform when we run the release policy.
In the case scpsys.c include mt2701 headers.
That is not allowed when we only release mt6761 drivers.
This patch rename to v platform header to workaorund it.

MTK-Commit-Id: acdf01b529e8ec06c1251b29c8ba6cea46f4bc3b

Test: Build and Boot
Change-Id: I134f46e9031a8640e30c8845a6f0acd22b40eef9
CR-Id: ALPS04982554
Feature: Thermal Management
Signed-off-by: Juju Sung <juju.sung@mediatek.com>