[ALPS04846031] GPU: Fix load rgx.sh failed
[Detail]
This prebuilt TQ shader was loaded whenever any process needs to use it,
But it isnt every process has permission to load the firmware (vendor_file).
So let it be loaded when hal_init() in gralloc, together with rgx.fw.
Revert the user helper workaround used to load firmware.

[Support]
Ticket: 124904

MTK-Commit-Id: ecf14d84394b4fc9f3d9059ca527533d37c76aaf

Change-Id: Ibbe26fe9d956f860487d73ff711321838da62d37
Signed-off-by: Eric Chung <eric.chung@mediatek.com>
CR-Id: ALPS04846031
Feature: OpenGL|ES