[ALPS04865088] Kinibi 500: Fix IRQ line number setting
Fix setting for IRQ line number setting for SWd to NWd notification.
SWd and NWd will not be able to communicate if this setting is
incorrect and since on Mediatek platform it shifts 32, the option
we used on kernel-4.14(CONFIG_MTK_SYSIRQ) to detect this should be
changed to kernel-4.19(CONFIG_ARCH_MEDIATEK) instead.

MTK-Commit-Id: a9dfe33e41ed38132a228a54e54120165acd1167

Change-Id: Ib99dc85f43b52e30637801322f75998ff71ea74e
Signed-off-by: Pangyen Chen <pangyen.chen@mediatek.com>
CR-Id: ALPS04865088
Feature: Trustonic TEE (Trusted Execution Environment)