[ALPS06278381] net: split out functions related to registering inflight socket files
commit f4e65870e5cede5ca1ec0006b6c9803994e5f7b8 upstream.

We need this functionality for the io_uring file registration, but
we cannot rely on it since CONFIG_UNIX can be modular. Move the helpers
to a separate file, thats always builtin to the kernel if CONFIG_UNIX is
m/y.

No functional changes in this patch, just moving code around.

MTK-Commit-Id: 0172197d30453459d432ee91cdebcea7120990c9

Reviewed-by: Hannes Reinecke <hare@suse.com>
Acked-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
[ backported to older kernels to get access to unix_gc_lock - gregkh ]
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Feature: Google Security Patch
CR-Id: ALPS06278381
Change-Id: I8549e229d4df80bcff4d2c3fa4b4a8eeb5218ae5