[ALPS06511517] binder: use cred instead of task for selinux checks
Since binder was integrated with selinux, it has passed
struct task_struct associated with the binder_proc
to represent the source and target of transactions.
The conversion of task to SID was then done in the hook
implementations. It turns out that there are race conditions
which can result in an incorrect security context being used.

Fix by using the struct cred saved during binder_open and pass
it to the selinux subsystem.

MTK-Commit-Id: dc0f56b195ff4cd16b66bc8deecc56ab2b5c9d3b

Cc: stable@vger.kernel.org # 5.14 (need backport for earlier stables)
Fixes: 79af73079d75 ("Add security hooks to binder and implement the hooks for SELinux.")
Suggested-by: Jann Horn <jannh@google.com>
Signed-off-by: Todd Kjos <tkjos@google.com>
Acked-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>
Change-Id: Id7157515d2b08f11683aeb8ad9b8f1da075d34e7
Bug: 200688826
[ tkjos@ fixed minor conflict ]
Signed-off-by: Frankie Chang <frankie.chang@mediatek.com>
CR-Id: ALPS06511517
Feature: Google Security Patch