[ALPS07405919] GPU 1.15:Fix coverity issue
[Detail]
There is a potential use of uninitialised data when calling
PVRSRVRGXCreateTransferContextKM with a zero sized Framework command.
This is caused by a missing initialisation of the local sInfo
structure within PVRSRVRGXCreateTransferContextKM.
Symptoms are a kernel-exception which could lead to a platform crash.
[Solution]
The fix is to explicitly initialise the data structure to NULL so
that there is no longer any chance of using uninitialised data.

[Support]
Ticket:133635
PP133635.CL6277966.Fix.uninitialised.data.in.
PVRSRVRGXCreateTransferContextKM.patch

MTK-Commit-Id: 2dbaef8ca8b91a38bab0dc44276bed1eff5ad883

CR-Id: ALPS07405919
Feature: IMG GPU DDK driver
Signed-off-by: Huan Feng <huan.feng@mediatek.com>
Change-Id: I730fadaf17f62552f8e3d656d52a0ec245114941