[ALPS07405919] GPU 1.15:Fix coverity issue
[Detail]
There is a potential use of uninitialised data when calling
PVRSRVRGXCreateTransferContextKM with a zero sized Framework command.
This is caused by a missing initialisation of the local sInfo
structure within PVRSRVRGXCreateTransferContextKM.
Symptoms are a kernel-exception which could lead to a platform crash.
[Solution]
The fix is to explicitly initialise the data structure to NULL so
that there is no longer any chance of using uninitialised data.

[Support]
Ticket:133635
PP133635.CL6277966.Fix.uninitialised.data.in.
PVRSRVRGXCreateTransferContextKM.patch

MTK-Commit-Id: 15ad40cf904be39d3085583f5e8b9d80e7622069

CR-Id: ALPS07405919
Feature: IMG GPU DDK driver
Signed-off-by: Huan Feng <huan.feng@mediatek.com>
Change-Id: Id26529bfe1d635cfc17beed88de116d4562d8300