[ALPS07560723] touch:fix oob issue
When reading /dev/NVTSPI, nvt_flash_read is called. And the “buff” and “count” parameters are user controlled. The “str” is copied from “buff”. Then, “((str[0] & 0x7F) << 8) | str[1]” is used to generate a length to copy “str+2” to “buf”, which is allocated with size “count”. When the user uses a small “count” (e.g. 68) and a large enough “str[0]” and “str[1]” (e.g "0xff 0xff" or "0x7F 0xff”), there is a oob write to “buf”.

MTK-Commit-Id: 93ec5674d76453323973d3ef88996c7618afac35

Signed-off-by: ot903067 <ot_he.liu@mediatek.com>
CR-Id: ALPS07560723
Feature: touch_driver
Change-Id: I3036e91d1361acd545b23962259b7713ae8790c7