[ALPS07770895] fbcon: Disallow setting font bigger than screen size
commit 65a01e601dbba8b7a51a2677811f70f783766682 upstream.

Prevent that users set a font size which is bigger than the physical screen. Its unlikely this may happen (because screens are usually much larger than the fonts and each font char is limited to 32x32 pixels), but it may happen on smaller screens/LCD displays.

Signed-off-by: Helge Deller <deller@gmx.de>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Reviewed-by: Geert Uytterhoeven <geert@linux-m68k.org>
Cc: stable@vger.kernel.org # v4.14+
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

MTK-Commit-Id: 34cf18adeb48b9d6ae437f6072721cbf71742c06

Signed-off-by: Hongjiang He <Hongjiang.He@mediatek.com>
CR-Id: ALPS07770895
Feature: Google Security Patch
Change-Id: Ib2ec6a64922800eb139006952374f105847187ac