[N12][gki][android14-6.1-2024-01_r15] [1/5]ANDROID: arm64: Call fixup_exception() within do_sea()
ROTHKO-22403

path_lookupat() is capable of safely reading unampped VAs. If an
unmapped VA is read whilst the function is being called, the resulting
page fault will get re-directed to __do_page_fault(), which will call
fixup_exception() to handle the aforementioned unmapped VA read.

Now, for an OS running in a VM, let's say that memory was still mapped
at S1 but lent to another VM (i.e. unmapped at S2 for the given VM).
The reading of an unmapped VA in path_lookupat() still needs to be
handled. For hypervisors that inject an abort leading to a do_sea()
call, call fixup_exception() from do_sea() if
trace_android_vh_try_fixup_sea() indicates that we can do so.

Bug: 320358381
Bug: 336464338

Change-Id: I197f28ee9be3b681a7e584295f2e1b48b11721a5
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>